Modifies closeResource test to include features for opening and switching between resources
